requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dateutil==2.9.0.post0
requests==2.32.3
jinja2==3.1.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
structlog==23.2.0
openai==1.3.0
numpy==1.24.3
scikit-learn==1.3.0
alembic==1.12.1
redis==5.0.1
hiredis==2.2.3
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0
PyPDF2==3.0.1
python-docx==1.1.0
